requirements.txt
google-cloud-compute==2.11.0
google-cloud-storage==2.13.0
paramiko==3.4.0
requests==2.31.0
